pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "polyedge"
version = "0.1.0"
description = "AI-powered Polymarket trading bot"
requires-python = ">=3.11"
dependencies = [
    "py-clob-client>=0.17.0",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "asyncpg>=0.29.0",
    "aiohttp>=3.9.0",
    "anthropic>=0.40.0",
    "openai>=1.50.0",
    "click>=8.1.0",
    "rich>=13.0.0",
    "pyyaml>=6.0",
    "eth-account>=0.11.0",
    "web3>=6.0.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-mock>=3.12",
]

[project.scripts]
polyedge = "polyedge.cli:cli"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
